requirements.txt
flask
pandas
openpyxl
scikit-learn
python-levenshtein
werkzeug